Restructured files to align with already-exisiting layout. No longer build using CMake.
